pyproject.toml
[tool.poetry]
name = "saverbot"
version = "0.1.0"
description = ""
authors = ["midnight <ivanator666@gmail.com>"]
license = "GPL v3"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.32.3"
python-telegram-bot = "^21.6"
python-dotenv = "^1.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
